requirements.txt
Flask
Flask-bootstrap4
pymongo